Finished the everything, just testing left.
Everything is done.
The robot should move forwards 93.3 inches in autonomous and should be
able to be controlled with the joystick in teleop.
The winch should also be moved in teleop when button 1 is pressed as
well as buttons 7-12 controlling the teleop speed.
I added a small amount of course autocorrect to the autonomous code, it
needs to be tested to make sure that it isn't over correcting.

TODO:  Add different starting positions and make it so the robot moves
in more preset paths if certain conditions are met.  Maybe use the
rangefinder to look forwards and see if the centre hexagon thing is
there or turn it on an angle and have it look for the walls.